+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.0.0-alpha.10](https://github.com/zama-ai/sdk/compare/v1.0.0-alpha.9...v1.0.0-alpha.10) (2026-03-06)
+
+### Features
+
+- add cleartext fhEVM feature ([#51](https://github.com/zama-ai/sdk/issues/51)) ([f951673](https://github.com/zama-ai/sdk/commit/f951673a86e975c06dcd41527c5401c365a7d2d5))
+
+## [1.0.0-alpha.9](https://github.com/zama-ai/sdk/compare/v1.0.0-alpha.8...v1.0.0-alpha.9) (2026-03-06)
+
+### Bug Fixes
+
+- **release:** remove registry-url to fix npm trusted publishing ([aa515ee](https://github.com/zama-ai/sdk/commit/aa515ee4329e83347049513aa1435f89d8a13df8))
+
 ## [1.0.0-alpha.8](https://github.com/zama-ai/sdk/compare/v1.0.0-alpha.7...v1.0.0-alpha.8) (2026-03-05)
 
 ### Features